Add button to start scan QR
